refactor: Separate providers and tests by contest category (#3169)
